refactor(look): render photo view as timeline + open lightbox
Fold the home and photo routes into one shared TimelinePage: the URL's
photoId opens the lightbox over the timeline, and an unknown id (bad
link, deleted photo, load error) leaves it closed on the grid instead of
a blank full-screen dialog — matching the album route's not-found parity.
